+++ Pulse.h
@@ -13,7 +13,7 @@ private:
 	bool fade;
 	bool mirrored;
 
-	float fadeRate;
+	uint8_t fadeRate;
 	uint8_t colorIndex;
 	uint8_t changeRate;
